[Fix] ScholorshipBoxView 디테일 수정
